修改全屏不传订单数

+++ follow-order-followcom/src/main/java/net/maku/followcom/vo/AccountCacheVO.java
@@ -50,7 +50,7 @@ public class AccountCacheVO {
     private BigDecimal equity;
     //
     @JsonProperty(value = "Credit")
-    private BigDecimal credit;
+    private Double credit;
     //
     @JsonProperty(value = "Profit")
     private Double profit;
